refactor: 메인화면 프래그먼트 패키지 구조 수정
