+++ app/src/androidTest/java/woowacourse/movie/view/main/MoviesFragmentTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.view.main
 
 import androidx.fragment.app.testing.launchFragmentInContainer
 import androidx.test.espresso.Espresso.onView
@@ -7,6 +7,7 @@ import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import org.junit.jupiter.api.Test
 import woowacourse.movie.R
+import woowacourse.movie.view.main.home.MoviesFragment
 
 class MoviesFragmentTest {
     @Test
