+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.movie.MoviesActivity"
+            android:name=".view.main.MoviesActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -30,7 +30,7 @@
         <activity
             android:name=".view.reservation.complete.ReservationCompleteActivity"
             android:exported="false"
-            android:parentActivityName=".view.movie.MoviesActivity" />
+            android:parentActivityName=".view.main.MoviesActivity" />
     </application>
 
 </manifest>
