+++ app/src/main/java/woowacourse/movie/view/main/home/MovieContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.view.main.home
 
 import woowacourse.movie.model.movie.Movie
