+++ app/src/main/java/woowacourse/movie/view/main/home/MoviePresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.view.main.home
 
 import woowacourse.movie.model.database.MovieDao
