+++ app/src/main/java/woowacourse/movie/view/main/home/MoviesFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.view.main.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,8 +8,8 @@ import androidx.fragment.app.Fragment
 import woowacourse.movie.databinding.FragmentMoviesBinding
 import woowacourse.movie.model.database.DummyMovieDao
 import woowacourse.movie.model.movie.Movie
-import woowacourse.movie.util.ExceptionMessages.FRAGMENT_BINDING_STATE_EXCEPTION
-import woowacourse.movie.view.movie.adapter.MovieAdapter
+import woowacourse.movie.util.ExceptionMessages
+import woowacourse.movie.view.main.home.adapter.MovieAdapter
 import woowacourse.movie.view.theater.TheaterBottomSheetDialogFragment
 
 class MoviesFragment :
@@ -18,7 +18,9 @@ class MoviesFragment :
     @Suppress("ktlint:standard:backing-property-naming")
     private var _binding: FragmentMoviesBinding? = null
     private val binding: FragmentMoviesBinding
-        get() = _binding ?: throw IllegalStateException(FRAGMENT_BINDING_STATE_EXCEPTION)
+        get() =
+            _binding
+                ?: throw IllegalStateException(ExceptionMessages.FRAGMENT_BINDING_STATE_EXCEPTION)
     private val presenter: MoviePresenter by lazy { MoviePresenter(this, DummyMovieDao) }
     private val moviesAdapter: MovieAdapter by lazy {
         MovieAdapter { movie ->
@@ -49,7 +51,7 @@ class MoviesFragment :
     }
 
     override fun showBottomSheetDialog(movie: Movie) {
-        val bottomSheet = TheaterBottomSheetDialogFragment.newInstance(movie)
+        val bottomSheet = TheaterBottomSheetDialogFragment.Companion.newInstance(movie)
         bottomSheet.show(parentFragmentManager, BOTTOM_SHEET_TAG)
     }
