+++ app/src/main/java/woowacourse/movie/view/main/reservationlist/ReservationListFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.main.reservationlist
 
 import android.os.Bundle
 import android.view.LayoutInflater
