+++ app/src/main/res/layout/fragment_movies.xml
@@ -7,14 +7,14 @@
 
         <variable
             name="fragmentMovies"
-            type="woowacourse.movie.view.movie.MoviesFragment" />
+            type="woowacourse.movie.view.main.home.MoviesFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.movie.MoviesActivity">
+        tools:context=".view.main.MoviesActivity">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_movies"
