+++ app/src/main/res/layout/fragment_reservation_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ReservationListFragment">
+    tools:context=".view.main.setting.SettingFragment">
 
     <TextView
         android:layout_width="match_parent"
