+++ app/src/main/res/layout/fragment_setting.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SettingFragment">
+    tools:context=".view.main.setting.SettingFragment">
 
     <TextView
         android:layout_width="match_parent"
